fix test_expr_consists_of_constants
